requirements.txt
streamlit==1.31.1
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
scikit-learn==1.3.2
scipy==1.11.4
plotly==5.18.0
python-dotenv==1.0.0
openai==0.28.1
tqdm==4.66.1
